# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationMethod": {
        "description": "Métodos de avaliação disponíveis na fatia representativa.",
        "enum": [
          "deterministic_match",
          "programmatic_check",
          "llm_as_judge"
        ],
        "title": "EvaluationMethod",
        "type": "string"
      },
...
      "EvaluationSession": {
        "description": "Snapshot durável do progresso de uma avaliação live.",
        "properties": {
          "case_statuses": {
            "additionalProperties": {
              "enum": [
                "stable_pass",
                "stable_fail",
                "unstable",
                "inconclusive"
              ],
              "type": "string"
            },
            "title": "Case Statuses",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_case_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Case Id"
          },
          "current_trial": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Trial"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "event_cursor": {
            "default": 0,
            "title": "Event Cursor",
            "type": "integer"
          },
          "final_verdict": {
            "$ref": "#/components/schemas/FinalVerdict",
            "default": "inconclusive"
          },
          "request": {
            "$ref": "#/components/schemas/SessionRequest"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunVerdict"
            },
            "title": "Runs",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "cancelled",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "created_at",
          "updated_at",
          "status",
          "request"
        ],
        "title": "EvaluationSession",
        "type": "object"
      },
      "FinalVerdict": {
        "description": "Resultado normativo de uma tentativa.",
        "enum": [
          "pass",
          "fail",
          "inconclusive"
        ],
        "title": "FinalVerdict",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "description": "Solicitação de execução para um caso conhecido do dataset.",
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateInput"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/EvaluationMethod"
            },
            "minItems": 1,
            "title": "Methods",
            "type": "array"
          }
        },
        "required": [
          "case_id",
          "methods",
          "candidate"
        ],
        "title": "RunRequest",
        "type": "object"
      },
      "RunVerdict": {
        "description": "Veredito persistido e renderizado integralmente pela interface.",
        "properties": {
          "agent_input": {
            "title": "Agent Input",
            "type": "string"
          },
          "agent_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Model"
          },
          "agent_output": {
            "title": "Agent Output",
            "type": "string"
          },
          "agent_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Provider"
          },
          "candidate_origin": {
            "enum": [
              "supplied_trace",
              "live_model"
            ],
            "title": "Candidate Origin",
            "type": "string"
          },
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationResult"
            },
            "title": "Evaluations",
            "type": "array"
          },
          "final_verdict": {
            "$ref": "#/components/schemas/FinalVerdict"
          },
          "judge_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Model"
          },
          "judge_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Provider"
          },
          "metrics": {
            "$ref": "#/components/schemas/RunMetrics"
          },
          "retrieved_context": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Retrieved Context",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "run_status": {
            "enum": [
              "completed",
              "partial",
              "failed"
            ],
            "title": "Run Status",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "trial_index": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Index"
          }
        },
        "required": [
          "run_id",
          "case_id",
          "created_at",
          "run_status",
          "final_verdict",
          "candidate_origin",
          "agent_input",
          "agent_output",
          "tool_calls",
          "evaluations",
          "metrics"
        ],
        "title": "RunVerdict",
        "type": "object"
      },
...
      "SessionRequest": {
        "description": "Cria uma execução live de um caso ou suíte fechada.",
        "properties": {
          "case_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 24,
            "minItems": 1,
            "title": "Case Ids",
            "type": "array"
          },
          "concurrency": {
            "default": 2,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Concurrency",
            "type": "integer"
          },
          "dataset_id": {
            "default": "builtin-v2",
            "title": "Dataset Id",
            "type": "string"
          },
          "judge_role": {
            "default": "judge_alt",
            "enum": [
              "judge",
              "judge_alt"
            ],
            "title": "Judge Role",
            "type": "string"
          },
          "methods": {
            "items": {
              "$ref": "#/components/schemas/EvaluationMethod"
            },
            "title": "Methods",
            "type": "array"
          },
          "mode": {
            "default": "single",
            "enum": [
              "single",
              "suite"
            ],
            "title": "Mode",
            "type": "string"
          },
          "pricing_profiles": {
            "additionalProperties": {
              "enum": [
                "standard",
                "free",
                "paid_standard"
              ],
              "type": "string"
            },
            "title": "Pricing Profiles",
            "type": "object"
          },
          "trials": {
            "const": 3,
            "default": 3,
            "title": "Trials",
            "type": "integer"
          }
        },
        "required": [
          "case_ids"
        ],
        "title": "SessionRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Avaliação de agentes com racional auditável.",
    "title": "agent-eval-harness",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/eval/run": {
      "post": {
        "operationId": "evaluate_api_eval_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunVerdict"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate"
      }
    },
    "/api/eval/runs": {
      "get": {
        "operationId": "list_runs_api_eval_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunVerdict"
                  },
                  "title": "Response List Runs Api Eval Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs"
      }
    },
    "/api/eval/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_eval_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/api/eval/sessions": {
      "get": {
        "operationId": "list_sessions_api_eval_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationSession"
                  },
                  "title": "Response List Sessions Api Eval Sessions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions"
      },
      "post": {
        "description": "Persiste e inicia uma avaliação real com três tentativas.",
        "operationId": "create_session_api_eval_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session"
      }
    },
    "/api/eval/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_api_eval_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session"
      }
    },
    "/api/eval/sessions/{session_id}/cancel": {
      "post": {
        "operationId": "cancel_session_api_eval_sessions__session_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Session"
      }
    },
    "/api/eval/sessions/{session_id}/events": {
      "get": {
        "description": "Transmite marcos persistidos e tokens efêmeros no formato SSE.",
        "operationId": "session_events_api_eval_sessions__session_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session Events"
      }
    },
...
  }
}